Rename output profiles
